API

Readability.ari — Function

ari(text::String)

Returns the Automated Readability Index (ARI) of text.

Argument

text::String: The text to analyze.

Returns

Int: The ARI score.

Reference

ARI

Readability.characters — Function

characters(text::String)

Returns the number of characters in text.

Readability.characters_per_word — Function

characters_per_word(text::String)

Returns the number of characters per word or the average word length in text.

Readability.coleman_liau — Function

coleman_liau(text::String)

Returns the Coleman-Liau Index of text.

CLI = 0.0588 * L - 0.296 * S - 15.8

L = number of characters per 100 words

S = number of sentences per 100 words

Argument

text::String: The text to analyze.

Returns

Number: The Coleman-Liau Index.

Readability.complex_words — Function

complex_words(text::String)

Returns the number of complex words (words with 3 or more syllables and not ending in "es", "ed", or "ing") in text.

Readability.dale_chall — Function

dale_chall(text::String)::Number

Returns the Dale-Chall readability score of text.

Readability.difficult_words — Function

difficult_words(text::String, word_list::String)::Int

Returns the number of words that are not in the specified word_list (either "dale-chall" or "spache") in text.

Readability.flesch_kincaid_grade_level — Function

flesch_kincaid_grade_level(text::String)

Returns the Flesch-Kincaid grade level of text.

Readability.flesch_reading_ease_score — Function

flesch_reading_ease_score(text::String)

Returns the Flesch Reading Ease Score of text.

Readability.gunning_fog — Function

gunning_fog(text::String)

Returns the Gunning Fog index of text.

Readability.lines — Function

lines(text::String)

Returns the number of lines text.

Readability.paragraphs — Function

paragraphs(text::String)

Returns the number of paragraphs in text.

Readability.polysyllabic_words — Function

polysyllabic_words(text::String)

Returns the number of words with 3 or more syllables in text.

Readability.reading_time — Function

reading_time(text::String; wpm::Number=238)::Number

Returns the reading time of text in seconds.

Arguments

text::String: The text to analyze.

wpm::Number: The words per minute.

Returns

Float64: The reading time in seconds.

Readability.sentences — Function

sentences(text::String)

Returns the number of sentences in text.

Readability.sentences_per_paragraph — Function

sentences_per_paragraph(text::String)

Returns the number of sentences per paragraph or the average paragraph length in text.

Readability.smog — Function

smog(text::String)::Number

Returns the SMOG Index of text.

Readability.spache — Function

spache(text::String)

Returns the Spache readability score of text.

Readability.speaking_time — Function

speaking_time(text::String; wpm::Number=183)::Number

Returns the speaking time of text in seconds.

Arguments

text::String: The text to analyze.

wpm::Number: The words per minute.

Returns

Float64: The speaking time in seconds.

Readability.syllables — Function

syllables(text::String)

Returns the number of syllables in text.

Readability.syllables_per_word — Function

syllables_per_word(text::String)

Returns the number of syllables per word or the average word length in text.

Readability.words — Function

words(text::String)

Returns the number of words in text.

Readability.words_per_sentence — Function

words_per_sentence(text::String)

Returns the number of words per sentence or the sentence length in text.